generate_qr_code/requirements.txt
functions-framework==3.*
qrcode
pillow
flask
google-cloud-storage
